Created error layouts in HistoryItemFragment

+++ presentation/src/main/java/kz/kamadi/yandextranslate/ui/history/HistoryItemFragment.java
@@ -6,15 +6,20 @@ import android.os.Bundle;
 import android.os.Handler;
 import android.support.annotation.Nullable;
 import android.support.v7.app.AlertDialog;
+import android.support.v7.widget.DividerItemDecoration;
 import android.support.v7.widget.LinearLayoutManager;
 import android.support.v7.widget.RecyclerView;
 import android.text.Editable;
 import android.text.TextWatcher;
+import android.util.Log;
 import android.view.MotionEvent;
 import android.view.View;
 import android.view.inputmethod.InputMethodManager;
 import android.widget.ImageButton;
+import android.widget.ImageView;
+import android.widget.LinearLayout;
 import android.widget.RelativeLayout;
+import android.widget.TextView;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -28,12 +33,14 @@ import kz.kamadi.yandextranslate.R;
 import kz.kamadi.yandextranslate.data.entity.History;
 import kz.kamadi.yandextranslate.presenter.HistoryItemPresenter;
 import kz.kamadi.yandextranslate.ui.base.BaseFragment;
+import kz.kamadi.yandextranslate.ui.listener.EndlessRecyclerViewScrollListener;
+import kz.kamadi.yandextranslate.ui.listener.OnLoadMoreListener;
 import kz.kamadi.yandextranslate.ui.listener.OnPageVisibleListener;
 import kz.kamadi.yandextranslate.ui.widgets.TranslateEditText;
 
 import static android.content.Context.INPUT_METHOD_SERVICE;
 
-public class HistoryItemFragment extends BaseFragment implements HistoryItemView, OnPageVisibleListener, HistoryAdapter.OnHistoryActionListener, TextWatcher, TranslateEditText.EditTextImeBackListener, View.OnTouchListener {
+public class HistoryItemFragment extends BaseFragment implements HistoryItemView, OnPageVisibleListener, OnLoadMoreListener, HistoryAdapter.OnHistoryActionListener, TextWatcher, TranslateEditText.EditTextImeBackListener, View.OnTouchListener {
     private static final String FAVOURITE = "kz.kamadi.yandextranslate.ui.history.HistoryItemFragment.FAVOURITE";
     @BindView(R.id.container)
     RelativeLayout container;
@@ -45,6 +52,12 @@ public class HistoryItemFragment extends BaseFragment implements HistoryItemView
     RelativeLayout searchLayout;
     @BindView(R.id.clear_image_button)
     ImageButton clearImageButton;
+    @BindView(R.id.error_layout)
+    LinearLayout errorLayout;
+    @BindView(R.id.error_icon)
+    ImageView errorIcon;
+    @BindView(R.id.error_message)
+    TextView errorMessage;
     @BindDrawable(R.drawable.border_bottom)
     Drawable borderBottom;
     @BindDrawable(R.drawable.border_bottom_yellow)
@@ -52,15 +65,16 @@ public class HistoryItemFragment extends BaseFragment implements HistoryItemView
     @Inject
     HistoryItemPresenter presenter;
 
-    private List<History> histories;
+    private List<History> histories = new ArrayList<>();
     private HistoryAdapter adapter;
     private int offset = 0;
-    private int limit = 25;
+    private int limit = 15;
     private boolean isFavourite;
     private Handler handler = new Handler();
     private boolean isKeyboardOpen = false;
     private boolean isSearching = false;
     private DeleteAllButtonVisibilityListener deleteAllButtonVisibilityListener;
+    private EndlessRecyclerViewScrollListener scrollListener;
 
 
     public static HistoryItemFragment newInstance(boolean isFavourite) {
@@ -93,20 +107,28 @@ public class HistoryItemFragment extends BaseFragment implements HistoryItemView
     @Override
     public void onViewCreated(View view, @Nullable Bundle savedInstanceState) {
         super.onViewCreated(view, savedInstanceState);
-        if (getParentFragment() instanceof DeleteAllButtonVisibilityListener){
+        if (getParentFragment() instanceof DeleteAllButtonVisibilityListener) {
             deleteAllButtonVisibilityListener = (DeleteAllButtonVisibilityListener) getParentFragment();
         }
-        if (isFavourite){
+        if (isFavourite) {
             searchEditText.setHint(R.string.search_favourites);
-        }else {
+            errorMessage.setText(R.string.favourites_error_message);
+            errorIcon.setImageResource(R.drawable.no_favourites);
+        } else {
             searchEditText.setHint(R.string.search_history);
+            errorMessage.setText(R.string.history_error_message);
+            errorIcon.setImageResource(R.drawable.no_history);
         }
         searchEditText.setOnTouchListener(this);
         searchEditText.addTextChangedListener(this);
         searchEditText.setOnEditTextImeBackListener(this);
         adapter = new HistoryAdapter(context, this);
         recyclerView.setAdapter(adapter);
-        recyclerView.setLayoutManager(new LinearLayoutManager(getActivity()));
+        LinearLayoutManager layoutManager = new LinearLayoutManager(getActivity());
+        recyclerView.setLayoutManager(layoutManager);
+        recyclerView.addItemDecoration(new DividerItemDecoration(context, layoutManager.getOrientation()));
+        scrollListener = new EndlessRecyclerViewScrollListener(layoutManager, this);
+        recyclerView.addOnScrollListener(scrollListener);
     }
 
 
@@ -115,7 +137,12 @@ public class HistoryItemFragment extends BaseFragment implements HistoryItemView
         searchEditText.setText("");
         getHistories();
         if (!isKeyboardOpen) {
-            openKeyboard();
+            handler.post(() -> {
+                InputMethodManager inputMethodManager = (InputMethodManager) getActivity().getSystemService(INPUT_METHOD_SERVICE);
+                inputMethodManager.toggleSoftInputFromWindow(searchEditText.getApplicationWindowToken(), InputMethodManager.SHOW_FORCED, 0);
+                searchEditText.requestFocus();
+                isKeyboardOpen = true;
+            });
         }
         changeSearchLayout(true);
     }
@@ -142,20 +169,36 @@ public class HistoryItemFragment extends BaseFragment implements HistoryItemView
 
     @Override
     public void showHistories(List<History> histories) {
-        if (!isSearching && deleteAllButtonVisibilityListener!=null){
+        if (!isSearching && deleteAllButtonVisibilityListener != null && this.histories.isEmpty()) {
             deleteAllButtonVisibilityListener.setVisibility(!histories.isEmpty());
         }
-        if (this.histories == null) {
-            this.histories = histories;
-            adapter.setHistories(histories);
-        } else {
-            adapter.add(histories);
+        if (!histories.isEmpty()){
+            scrollListener.setLoading(false);
         }
+        if (!isSearching && histories.isEmpty() && this.histories.isEmpty()) {
+            updateErrorLayoutVisibility(true);
+        } else {
+            updateErrorLayoutVisibility(false);
+            if (this.histories.isEmpty()) {
+                this.histories = histories;
+                adapter.setHistories(histories);
+            } else {
+                adapter.add(histories);
+            }
+        }
+        offset = this.histories.size();
     }
 
     @Override
     public void onHistoryDeleted(int position) {
         adapter.deleteItem(position);
+        if (this.histories.isEmpty()) {
+            updateErrorLayoutVisibility(true);
+            if (deleteAllButtonVisibilityListener != null) {
+                deleteAllButtonVisibilityListener.setVisibility(!histories.isEmpty());
+            }
+        }
+        offset = this.histories.size();
     }
 
     @Override
@@ -168,7 +211,7 @@ public class HistoryItemFragment extends BaseFragment implements HistoryItemView
     public void onPageHidden() {
         View view = getActivity().getCurrentFocus();
         if (view != null) {
-            InputMethodManager imm = (InputMethodManager)getActivity().getSystemService(Context.INPUT_METHOD_SERVICE);
+            InputMethodManager imm = (InputMethodManager) getActivity().getSystemService(Context.INPUT_METHOD_SERVICE);
             imm.hideSoftInputFromWindow(view.getWindowToken(), 0);
         }
         searchEditText.setText("");
@@ -187,7 +230,7 @@ public class HistoryItemFragment extends BaseFragment implements HistoryItemView
                 .setTitle(R.string.deleting)
                 .setMessage(R.string.deleting_message)
                 .setPositiveButton(android.R.string.yes, (dialog, which) -> {
-                    presenter.delete(history,position,isFavourite);
+                    presenter.delete(history, position, isFavourite);
                 })
                 .setNegativeButton(android.R.string.no, (dialog, which) -> {
 
@@ -215,7 +258,7 @@ public class HistoryItemFragment extends BaseFragment implements HistoryItemView
     @Override
     public void onTextChanged(CharSequence s, int start, int before, int count) {
         if (s.length() > 0) {
-            histories = null;
+            histories.clear();
             presenter.search(s.toString(), isFavourite);
         }
     }
@@ -242,6 +285,15 @@ public class HistoryItemFragment extends BaseFragment implements HistoryItemView
         return false;
     }
 
+    @Override
+    public void onLoadMore() {
+        if (!isSearching) {
+            scrollListener.setLoading(true);
+            Log.e("offset", offset + "");
+            presenter.getHistories(offset, limit, isFavourite);
+        }
+    }
+
     private void changeSearchLayout(boolean isActive) {
         if (isActive) {
             searchLayout.setBackground(borderBottomActive);
@@ -254,17 +306,20 @@ public class HistoryItemFragment extends BaseFragment implements HistoryItemView
         }
     }
 
-    private void openKeyboard() {
-        handler.post(() -> {
-            InputMethodManager inputMethodManager = (InputMethodManager) getActivity().getSystemService(INPUT_METHOD_SERVICE);
-            inputMethodManager.toggleSoftInputFromWindow(searchEditText.getApplicationWindowToken(), InputMethodManager.SHOW_FORCED, 0);
-            searchEditText.requestFocus();
-            isKeyboardOpen = true;
-        });
+    private void updateErrorLayoutVisibility(boolean isDisplay) {
+        if (isDisplay) {
+            searchLayout.setVisibility(View.GONE);
+            recyclerView.setVisibility(View.GONE);
+            errorLayout.setVisibility(View.VISIBLE);
+        } else {
+            searchLayout.setVisibility(View.VISIBLE);
+            recyclerView.setVisibility(View.VISIBLE);
+            errorLayout.setVisibility(View.GONE);
+        }
     }
 
-    public void getHistories() {
-        histories = null;
+    private void getHistories() {
+        histories.clear();
         offset = 0;
         adapter.setHistories(new ArrayList<>());
         presenter.getHistories(offset, limit, isFavourite);
